pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "celery[redis]>=5.6.2",
    "dj-database-url>=3.1.0",
    "django>=6.0.1",
    "django-admin-interface>=0.30.0",
    "django-admin-ordering>=0.20.0",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.12.0",
    "django-extensions>=4.1",
    "django-filter>=25.1",
    "meilisearch>=0.40.0",
    "django-nested-admin>=4.1.1",
    "django-tagulous>=2.1.0",
    "django-tinymce>=5.0.0",
    "djangorestframework>=3.16.0",
    "djiiif>=0.22",
    "djoser>=2.3.1",
    "psycopg[binary,pool]>=3.3.2",
    "uvicorn[standard]>=0.34.0",
    "python-dateutil>=2.9.0.post0",
    "pyyaml>=6.0.2",
    "whitenoise>=6.9.0",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "ipython>=9.3.0",
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
    "ruff>=0.14.14",
]

[tool.ruff]
line-length = 120
exclude = [
    "*/migrations/*",
]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

extend-select = ["E501"]

[tool.ruff.lint.isort]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
force-sort-within-sections = true
combine-as-imports = true

[tool.pytest.ini_options]
log_cli = false
minversion = "6.0"
log_cli_level = "INFO"
django_settings_module = "config.settings"
python_files = ["test_*.py", "*_test.py", "tests.py"]
addopts = "--ignore-glob=**/*/__init__.py"
